[TASK] implement config file and validation of Update and Delete

+++ Stock/Products.cs
@@ -20,6 +20,7 @@ namespace Stock
 
         private void Products_Load(object sender, EventArgs e)
         {
+            button1.Enabled = false;
             comboBox1.SelectedIndex = 0;
             LoadData();
         }
@@ -36,37 +37,41 @@ namespace Stock
 
         private void button2_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection("Data Source=DESKTOP-8TDSR33\\CSMOSQL;Initial Catalog=Stock;Integrated Security=True");
-            // Insert Login
-            con.Open();
-            bool status = false;
-            if(comboBox1.SelectedIndex == 0)
+            if (Validation())
             {
-                status = true;
-            }
-            else
-            {
-                status = false;
-            }
+                SqlConnection con = Connection.GetConnection();
+                // Insert Login
+                con.Open();
+                bool status = false;
+                if (comboBox1.SelectedIndex == 0)
+                {
+                    status = true;
+                }
+                else
+                {
+                    status = false;
+                }
 
-            var sqlQuery = "";
-            if (ifProductExist(con, textBox1.Text))
-            {
-                sqlQuery = @"UPDATE [dbo].[Products]
+                var sqlQuery = "";
+                if (ifProductExist(con, textBox1.Text))
+                {
+                    sqlQuery = @"UPDATE [dbo].[Products]
                                SET [ProductName] = '" + textBox2.Text + "',[ProductStatus] = '" + status + "' WHERE [ProductCode] = '" + textBox1.Text + "'";
-            }
-            else
-            {
-                sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductCode] ,[ProductName],[ProductStatus])
+
+                }
+                else
+                {
+                    sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductCode] ,[ProductName],[ProductStatus])
                             VALUES
                            ( '" + textBox1.Text + "','" + textBox2.Text + "','" + status + "')";
-            }
+                }
 
-            SqlCommand cmd = new SqlCommand(sqlQuery, con);
-            cmd.ExecuteNonQuery();
-            con.Close();
-            LoadData();
-            
+                SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                cmd.ExecuteNonQuery();
+                con.Close();
+                LoadData();
+                ResetRecords();
+            }
         }
         
         private bool ifProductExist(SqlConnection con, string productCode)
@@ -86,7 +91,7 @@ namespace Stock
 
         public void LoadData()
         {
-            SqlConnection con = new SqlConnection("Data Source=DESKTOP-8TDSR33\\CSMOSQL;Initial Catalog=Stock;Integrated Security=True");
+            SqlConnection con = Connection.GetConnection();
             // Reading data
             SqlDataAdapter sda = new SqlDataAdapter("SELECT * FROM [Stock].[dbo].[Products]", con);
             DataTable dt = new DataTable();
@@ -110,6 +115,8 @@ namespace Stock
 
         private void dataGridView1_MouseDoubleClick(object sender, MouseEventArgs e)
         {
+            button2.Text = "Update";
+            button1.Enabled = true;
             textBox1.Text = dataGridView1.SelectedRows[0].Cells[0].Value.ToString();
             textBox2.Text = dataGridView1.SelectedRows[0].Cells[1].Value.ToString();
             if (dataGridView1.SelectedRows[0].Cells[2].Value.ToString() == "Active" )
@@ -124,21 +131,47 @@ namespace Stock
 
         private void button1_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection("Data Source=DESKTOP-8TDSR33\\CSMOSQL;Initial Catalog=Stock;Integrated Security=True");
-            var sqlQuery = "";
-            if (ifProductExist(con, textBox1.Text))
+            if (Validation())
             {
-                con.Open();
-                sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
-                SqlCommand cmd = new SqlCommand(sqlQuery, con);
-                cmd.ExecuteNonQuery();
-                con.Close();
+                SqlConnection con = Connection.GetConnection();
+                var sqlQuery = "";
+                if (ifProductExist(con, textBox1.Text))
+                {
+                    con.Open();
+                    sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
+                    SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                    cmd.ExecuteNonQuery();
+                    con.Close();
+                }
+                else
+                {
+                    MessageBox.Show("Record Not Exist...!");
+                }
+                LoadData();
+                ResetRecords();
             }
-            else
+        }
+        private void ResetRecords()
+        {
+            textBox1.Clear();
+            textBox2.Clear();
+            comboBox1.SelectedIndex = -1;
+            button2.Text = "Add";
+            textBox1.Focus();
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            ResetRecords();
+        }
+        private bool Validation()
+        {
+            bool result = false;
+            if (!string.IsNullOrEmpty(textBox1.Text) && !string.IsNullOrEmpty(textBox2.Text) && comboBox1.SelectedIndex > -1 )
             {
-                MessageBox.Show("Record Not Exist...!");
+                result = true;
             }
-            LoadData();
+            return result;
         }
     }
 }
